Fix SampleJsonDataGenerator overflow for long min/max values (#1901)
HandleIntegerType used Convert.ToInt32 which throws OverflowException
when schema minimum/exclusiveMinimum values exceed int range (e.g.,
long.MinValue or values > ~2.1 billion). Changed to Convert.ToInt64
to support the full range of integer values in JSON Schema.

Fixes #1589

+++ src/NJsonSchema/Generation/SampleJsonDataGenerator.cs
@@ -151,15 +151,15 @@ namespace NJsonSchema.Generation
         {
             if (schema.ExclusiveMinimumRaw != null)
             {
-                return JToken.FromObject(Convert.ToInt32(schema.ExclusiveMinimumRaw, CultureInfo.InvariantCulture));
+                return JToken.FromObject(Convert.ToInt64(schema.ExclusiveMinimumRaw, CultureInfo.InvariantCulture));
             }
             else if (schema.ExclusiveMinimum != null)
             {
-                return JToken.FromObject(Convert.ToInt32(schema.ExclusiveMinimum, CultureInfo.InvariantCulture));
+                return JToken.FromObject(Convert.ToInt64(schema.ExclusiveMinimum, CultureInfo.InvariantCulture));
             }
             else if (schema.Minimum.HasValue)
             {
-                return Convert.ToInt32(schema.Minimum, CultureInfo.InvariantCulture);
+                return Convert.ToInt64(schema.Minimum, CultureInfo.InvariantCulture);
             }
             return JToken.FromObject(0);
         }
